refactor: 도메인 아키텍처 구조 개선 및 불필요한 파일 정리
