+++ api/lecture/index.ts
@@ -1,2 +1,3 @@
-export { getLectureList } from './getLectureList';
 export { getLectureInfo } from './getLectureInfo';
+export { getLectureListStatic } from './getLectureListStatic';
+export { getLectureRating } from './getLectureRating';
