+++ components/common/index.ts
@@ -2,4 +2,3 @@ export { default as StarRating } from './starRating/StarRating';
 export { default as Selector } from './Selector/Selector';
 export { default as Modal } from './modal/Modal';
 export { default as ReviewDetailModal } from './modal/ReviewDetailModal';
-export { default as LoginModal } from './modal/LoginModal';
