+++ domains/auth/client/components/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginModal } from './LoginModal';
