+++ domains/auth/client/hooks/useLoginMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
-import { LoginParams, LoginResponse } from '@/types/auth';
-import { postLogin } from '@/api';
+import { LoginParams, LoginResponse } from '@/domains/auth';
+import { postLogin } from '@/api/auth';
 
 export function useLoginMutation() {
   return useMutation<LoginResponse, Error, LoginParams>({
