+++ domains/auth/index.ts
@@ -1,0 +1,5 @@
+export { LoginModal } from './client/components';
+
+export { useLoginMutation } from './client/hooks';
+
+export type { LoginParams, LoginResponse, UserData, AuthState, ModalState } from './shared/types';
