+++ domains/lecture/client/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useLectureRating } from './useLectureRating';
