+++ domains/lecture/shared/types/lectureInfo.ts
