+++ domains/lecture/styles/LectureList.module.css
