--- types/index.ts
@@ -1,4 +1,0 @@
-export * from './auth';
-export * from './lecture';
-export * from './lectureInfo';
-export * from './review';
